requirements.txt

langchain-community>=0.0.20
langchain-core>=0.1.33
langchain-huggingface>=0.0.2
langchain-qdrant>=0.1.1
langchain-ollama>=0.1.0
langchain-text-splitters>=0.0.1
streamlit>=1.28.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0
python-dotenv>=1.0.0
ollama>=0.1.7